pyproject.toml
[project]
name = "hubstaff-backend-api"
version = "0.1.0"
description = "FastAPI backend service for Hubstaff clone"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.1",
    "boto3>=1.34.0",
    "email-validator>=2.1.0",
    "fastapi>=0.104.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "pyjwt>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
]

[dependency-groups]
dev = [
    "black>=23.11.0",
    "flake8>=6.1.0",
    "httpx>=0.25.2",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "pre-commit>=3.5.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
extend-ignore = "E203,W503"
exclude = ".git,__pycache__,docs/source/conf.py,old,build,dist,.venv"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
python_classes = "Test*"
python_functions = "test_*"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=app --cov-report=term-missing"
asyncio_mode = "auto"
